pyproject.toml
[tool.poetry]
name = "veritas"
version = "0.1.0"
description = "nautobot toolkit library"
authors = ["Veritas"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.0"
pynautobot = "^2.0.2"
loguru = "^0.7.2"
pytricia = "^1.0.2"
openpyxl = "^3.1.2"
psycopg2-binary = "^2.9.9"
pyzmq = "^25.1.2"
anytree = "^2.12.1"
boolean-parser = "^0.1.4"
netaddr = "^0.10.1"
ttp = "^0.9.5"
pydriller = "^2.6"
textfsm = "^1.1.3"
nornir = "^3.4.1"
nornir-inspect = "^1.0.3"
nornir-napalm = "^0.4.0"
nornir-netmiko = "^1.0.1"
nornir-scrapli = "^2023.7.30"
nornir-utils = "^0.2.0"
python-slugify = "^8.0.1"
python-benedict = "^0.33.1"
ntc-templates = "^4.1.0"
pika = "^1.3.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
myst-nb = {version = "^1.0.0", python = "^3.11"}
sphinx-autoapi = "^3.0.0"
sphinx-rtd-theme = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
